+++ bin/sslverify.php
@@ -5,21 +5,17 @@ if (!defined('DS')) {
     define('DS', DIRECTORY_SEPARATOR);
 }
 
-declare(ticks = 100);
-
 require_once(__DIR__ . DS . '..' . DS . 'examples' . DS . 'bootstrap.php');
 
 use Symfony\Component\Console\Application as BaseApplication;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\ArrayInput;
-
-use HRDNS\Examples\Socket\EchoServer;
+use HRDNS\Examples\SSL\Validator;
 
 class Application extends BaseApplication
 {
 
-    const APP_NAME = 'HRDNS Examples';
+    const APP_NAME = 'SSL Validator';
     const APP_VERSION = '0.0.1';
     const AUTHOR_NAME = 'Richard Huelsberg';
     const AUTHOR_EMAIL = 'rh@hrdns.de';
@@ -29,11 +25,23 @@ class Application extends BaseApplication
         parent::__construct(self::APP_NAME, self::APP_VERSION);
     }
 
+    protected function getCommandName(InputInterface $input)
+    {
+        return 'exmaple:sslvalidator';
+    }
+
     protected function getDefaultCommands()
     {
         $commands = parent::getDefaultCommands();
-        $commands[] = new EchoServer();
+        $commands[] = new Validator();
         return $commands;
+    }
+
+    public function getDefinition()
+    {
+        $inputDefinition = parent::getDefinition();
+        $inputDefinition->setArguments();
+        return $inputDefinition;
     }
 
     public function doRun(InputInterface $input, OutputInterface $output)
@@ -41,8 +49,9 @@ class Application extends BaseApplication
         if (!$input->hasParameterOption('--quiet')) {
             $output->write(
                 sprintf(
-                    "<info>%s</info> by <comment>%s</comment> <%s>\n\n",
+                    "<info>%s</info> <comment>v%s</comment> by <comment>%s</comment> <%s>\n\n",
                     self::APP_NAME,
+                    $this->getVersion(),
                     self::AUTHOR_NAME,
                     self::AUTHOR_EMAIL
                 )
@@ -50,11 +59,6 @@ class Application extends BaseApplication
         }
         if ($input->hasParameterOption('--version') || $input->hasParameterOption('-V')) {
             return 0;
-        }
-        if (!$input->getFirstArgument()) {
-            $input = new ArrayInput(
-                array ('list')
-            );
         }
         return parent::doRun($input, $output);
     }