+++ examples/bootstrap.php
@@ -2,7 +2,7 @@
 
 set_time_limit(0);
 error_reporting(E_ALL | E_STRICT | E_DEPRECATED);
-ini_set('memory_limit',-1);
+ini_set('memory_limit', -1);
 
 if (!defined('DS')) {
     define('DS', DIRECTORY_SEPARATOR);
@@ -28,4 +28,3 @@ if (!$loaded) {
     fwrite(STDERR, $msg);
     exit(1);
 }
-