+++ src/HRDNS/Socket/Client/TCPClient.php
@@ -6,7 +6,7 @@ class TCPClient extends Client
 {
 
     /**
-     * @return static
+     * @return self
      * @throws \Exception
      */
     public function connect()