+++ src/HRDNS/Socket/Server/TCPServer.php
@@ -10,7 +10,7 @@ abstract class TCPServer extends Server
 
     /**
      * @param int $maxClients
-     * @return static
+     * @return self
      */
     public function setMaxClients($maxClients)
     {
@@ -19,7 +19,7 @@ abstract class TCPServer extends Server
     }
 
     /**
-     * @return static
+     * @return self
      * @throws \Exception
      */
     public function bind()
@@ -75,7 +75,7 @@ abstract class TCPServer extends Server
 
     /**
      * @param integer $limit
-     * @return static
+     * @return self
      */
     public function listen($limit = -1)
     {
@@ -105,7 +105,7 @@ abstract class TCPServer extends Server
     /**
      * @param ServerClient $client
      * @param bool $closeByPeer
-     * @return static
+     * @return self
      * @SuppressWarnings(PHPMD.BooleanArgumentFlag)
      */
     public function disconnect(ServerClient $client, $closeByPeer = false)
@@ -140,7 +140,7 @@ abstract class TCPServer extends Server
     }
 
     /**
-     * @return static
+     * @return self
      */
     protected function workOnMasterSocket()
     {
@@ -159,7 +159,7 @@ abstract class TCPServer extends Server
 
     /**
      * @param resource[] $read
-     * @return static
+     * @return self
      */
     protected function workOnClientSockets(array $read = array ())
     {
@@ -183,7 +183,7 @@ abstract class TCPServer extends Server
 
     /**
      * @param ServerClient $client
-     * @return static
+     * @return self
      */
     protected function workOnClientSocket(ServerClient $client)
     {