+++ tests/bootstrap.php
@@ -1,0 +1,7 @@
+<?php
+
+if (!defined('DS')) {
+    define('DS', DIRECTORY_SEPARATOR);
+}
+
+require_once(__DIR__ . DS . '..' . DS . 'vendor' . DS . 'autoload.php');